requirements.txt
pytest
requests
python-dotenv
allure-pytest
Faker
pika
prometheus-client